renamed profile_manager2 to profile_manager
