+++ src/core/zowe/core_for_zowe_sdk/profile_manager.py
@@ -13,14 +13,16 @@ Copyright Contributors to the Zowe Project.
 import base64
 import os.path
 import re
-import warnings
-from typing import Union
 import sys
+import warnings
+from dataclasses import dataclass
+from typing import Union
 
 import jsonc
 
 from .constants import constants
 from .exceptions import ProfileNotFound, SecureProfileLoadFailed, SecureValuesNotFound
+from .profile_constants import GLOBAL_CONFIG_NAME, TEAM_CONFIG, USER_CONFIG
 
 HAS_KEYRING = True
 try:
@@ -28,82 +30,100 @@ try:
 except ImportError:
     HAS_KEYRING = False
 
+HOME = os.path.expanduser("~")
+GLOBAL_CONFIG_PATH = os.path.join(HOME, ".zowe", f"{GLOBAL_CONFIG_NAME}.config.json")
+CURRENT_DIR = os.getcwd()
 
-class ProfileManager:
+
+@dataclass
+class ConfigFile:
     """
-    Class used to represent a Zowe z/OSMF profile.
+    Class used to represent a single config file.
 
-    Description
-    -----------
-    This class is only used when there is already a Zowe z/OSMF profile created
-    and the user opted to use the profile instead of passing the credentials directly
-    in the object constructor.
-
-    Attributes
-    ----------
-    profile_name: str
-        Zowe z/OSMF profile name
+    Mainly it will have the following details :
+    1. Type ("User Config" or "Team Config")
+        User Configs override Team Configs.
+        User Configs are used to have personalised config details
+        that the user don't want to have in the Team Config.
+    2. Directory in which the file is located.
+    3. Name (excluding .config.json or .config.user.json)
+    4. Contents of the file.
+    4.1 Profiles
+    4.2 Defaults
+    5. Secure Properties associated with the file.
     """
 
-    def __init__(self, appname: str = "zowe"):
-        self._appname = appname
-
-        # config directory
-        self._config_dir = None
-        self._user_config_dir = None
-
-        # default config filenames
-        self._config_filename = f"{self._appname}.config.json"
-        self._user_config_filename = f"{self._appname}.user.config.json"
-
-        # exact config paths
-        self._config_filepath = None
-        self._user_config_filepath = None
+    type: str
+    name: str
+    location: Union[str, None] = None
+    profiles: Union[dict, None] = None
+    defaults: Union[dict, None] = None
+    secure_props: Union[dict, None] = None
 
     @property
-    def config_appname(self) -> str:
-        """Returns the app name"""
-        return self._appname
+    def filename(self) -> str:
+        if self.type == TEAM_CONFIG:
+            return f"{self.name}.config.json"
+
+        if self.type == USER_CONFIG:
+            return f"{self.name}.config.user.json"
+
+        return self.name
 
     @property
-    def config_dir(self) -> Union[str, None]:
-        """Returns the folder path to where the Zowe z/OSMF Team Profile Config files are located."""
-        return self._config_dir
+    def filepath(self) -> Union[str, None]:
+        if not self.location:
+            return None
 
-    @config_dir.setter
-    def config_dir(self, dirname: str) -> None:
-        """
-        Set directory/folder path to where Zowe z/OSMF Team Profile Config files are located
-        """
-        if os.path.isdir(dirname):
-            self._config_dir = dirname
+        print(self.filename, self.location)
+        return os.path.join(self.location, self.filename)
+
+    def init_from_file(self) -> dict:
+        if self.filepath is None:
+            self.autodiscover_config_dir()
+
+        if self.filepath is None:
+            raise FileNotFoundError(f"Could not find the file {self.filename}")
+
+        with open(self.filepath, encoding="UTF-8", mode="r") as fileobj:
+            profile_jsonc = jsonc.load(fileobj)
+
+        print(profile_jsonc)
+        self.profiles = profile_jsonc["profiles"]
+        self.defaults = profile_jsonc["defaults"]
+
+        self.load_secure_props()
+
+    def get_profile(
+        self, profile_name: Union[str, None], profile_type: Union[str, None]
+    ):
+        if self.profiles is None:
+            self.init_from_file()
+
+        if profile_name is None and profile_type is None:
+            raise ProfileNotFound(
+                "Could not find profile as both profile_name and profile_type is not set."
+            )
+
+        if profile_name is None:
+            profile_name = self.get_profilename_from_profiletype(
+                profile_type=profile_type
+            )
+
+        props: dict = self.load_profile_properties(profile_name=profile_name)
+
+        try:
+            base_profile = self.get_profilename_from_profiletype(profile_type="base")
+        except ProfileNotFound:
+            if self.type == TEAM_CONFIG:
+                warnings.warn(f"Base profile not found in {self.filepath}")
         else:
-            raise FileNotFoundError(f"given path {dirname} is not valid")
+            base_props = self.load_profile_properties(profile_name=base_profile)
+            props.update(base_props)
 
-    @property
-    def user_config_dir(self) -> Union[str, None]:
-        """Returns the folder path to where the Zowe z/OSMF User Profile Config files are located."""
-        return self._user_config_dir
+        return props
 
-    @user_config_dir.setter
-    def user_config_dir(self, dirname: str) -> None:
-        """Set directory/folder path to where Zowe z/OSMF User Profile Config files are located"""
-        if os.path.isdir(dirname):
-            self._user_config_dir = dirname
-        else:
-            raise FileNotFoundError(f"given path {dirname} is not valid")
-
-    @property
-    def config_filename(self) -> str:
-        """Return the filename for Zowe z/OSMF Team Profile Config"""
-        return self._config_filename
-
-    @property
-    def config_filepath(self) -> Union[str, None]:
-        """Get the full Zowe z/OSMF Team Config filepath"""
-        return self._config_filepath
-
-    def autodiscover_config_dir(filename: str) -> Union[str, None]:
+    def autodiscover_config_dir(self):
         """
         Autodiscover Zowe z/OSMF Team Config files by going up the path from
         current working directory
@@ -112,11 +132,11 @@ class ProfileManager:
         Else, it returns None
         """
 
-        current_dir = os.getcwd()
+        current_dir = CURRENT_DIR
         config_dir = None
 
         while config_dir is None:
-            path = os.path.join(current_dir, filename)
+            path = os.path.join(current_dir, self.filename)
 
             if os.path.isfile(path):
                 config_dir = current_dir
@@ -127,44 +147,9 @@ class ProfileManager:
 
             current_dir = os.path.dirname(current_dir)
 
-        return config_dir
+        self.location = config_dir
 
-    def autodiscover_user_config_dir(self, filename: str) -> Union[str, None]:
-        """
-        Autodiscover Zowe z/OSMF User Profile Config files by
-        a. Try going up the path from current working directory
-        b. Trying to load from config_dir, if not found then
-        """
-
-        # try checking in current working directory or go up
-        current_dir = os.getcwd()
-        user_config_dir = None
-
-        while user_config_dir is None:
-            path = os.path.join(current_dir, filename)
-
-            if os.path.isfile(path):
-                user_config_dir = current_dir
-                return user_config_dir
-
-            # check if have arrived at the root directory
-            if current_dir == os.path.dirname(current_dir):
-                break
-
-            current_dir = os.path.dirname(current_dir)
-
-        # try checking in config dir
-        path = os.path.join(self._config_dir, filename)
-
-        if os.path.isfile(path):
-            user_config_dir = self._config_dir
-            return user_config_dir
-
-        return user_config_dir
-
-    def get_profilename_from_profiletype(
-        self, profile_jsonc: dict, profile_type: str
-    ) -> str:
+    def get_profilename_from_profiletype(self, profile_type: str) -> str:
         """
         Return exact profilename of the profile to load from the mentioned type
 
@@ -174,14 +159,14 @@ class ProfileManager:
         try:
             # try to get the profilename from defaults
             try:
-                profilename = profile_jsonc["defaults"][profile_type]
+                profilename = self.defaults[profile_type]
             except KeyError:
                 warnings.warn("Given profile type has no default profilename")
             else:
                 return profilename
 
             # iterate through the profiles and check if profile is found
-            for (key, value) in profile_jsonc["profiles"].items():
+            for (key, value) in self.profiles.items():
                 try:
                     temp_profile_type = value["type"]
                     if profile_type == temp_profile_type:
@@ -197,24 +182,20 @@ class ProfileManager:
         except ProfileNotFound as exc:
             raise exc
 
-    def load_profile_properties(self, profile_jsonc: dict, profile_name: str) -> dict:
+    def load_profile_properties(self, profile_name: str) -> dict:
         """
         Load exact profile properties (without prepopulated fields from base profile)
         from the profile dict and populate fields from the secure credentials storage
         """
         try:
-            props = profile_jsonc["profiles"][profile_name]["properties"]
-        except KeyError as exc:
-            raise ProfileNotFound(
-                profile_name=profile_name, error_msg=str(exc)
-            ) from exc
+            props = self.profiles[profile_name]["properties"]
+        except Exception as exc:
+            return {}
 
-        secure_props = self.load_credentials()
-
-        secure_fields: list = profile_jsonc["profiles"][profile_name].get("secure", [])
+        secure_fields: list = self.profiles[profile_name].get("secure", [])
 
         # load properties with key as profile.{profile_name}.properties.{*}
-        for (key, value) in secure_props.items():
+        for (key, value) in self.secure_props.items():
             if re.match("profiles\\." + profile_name + "\\.properties\\.[a-z]+", key):
                 property_name = key.split(".")[3]
                 if property_name in secure_fields:
@@ -226,145 +207,14 @@ class ProfileManager:
 
         return props
 
-    def load_base_profile_properties(self, profile_jsonc: dict) -> dict:
+    def load_secure_props(self) -> None:
         """
-        Load base profile
+        load secure_props stored for the given config
+
+        if keyring is not initialized, set empty value
         """
-        base_props: dict = {}
-
-        try:
-            base_profile_name = self.get_profilename_from_profiletype(
-                profile_jsonc=profile_jsonc, profile_type="base"
-            )
-            base_props = profile_jsonc["profiles"][base_profile_name].get(
-                "properties", {}
-            )
-        except KeyError as exc:
-            raise ProfileNotFound(
-                profile_name=base_profile_name, error_msg=str(exc)
-            ) from exc
-
-        secure_props = self.load_credentials()
-
-        secure_fields: list = profile_jsonc["profiles"][base_profile_name].get(
-            "secure", []
-        )
-
-        # load properties with key as profile.{profile_name}.properties.{*}
-        for (key, value) in secure_props.items():
-            if re.match(
-                "profiles\\." + base_profile_name + "\\.properties\\.[a-z]+", key
-            ):
-                property_name = key.split(".")[3]
-                if property_name in secure_fields:
-                    base_props[property_name] = value
-                    secure_fields.remove(property_name)
-
-        if len(secure_fields) > 0:
-            raise SecureValuesNotFound(secure_fields)
-
-        return base_props
-
-    def load_from_file(
-        self,
-        filepath: str,
-        profile_name: Union[str, None] = None,
-        profile_type: Union[str, None] = None,
-    ) -> dict:
-        with open(filepath, encoding="UTF-8", mode="r") as fileobj:
-            profile_jsonc = jsonc.load(fileobj)
-
-        # load profile
-        if profile_type is None and profile_name is None:
-            raise ValueError(
-                "Both profile_type and profile_name cannot be empty at the same time."
-            )
-
-        service_profile = {}
-
-        # load base profile
-        base_profile: dict = self.load_base_profile_properties(
-            profile_jsonc=profile_jsonc
-        )
-        service_profile.update(base_profile)
-
-        # load given profile
-        if profile_name is None:
-            profile_name = self.get_profilename_from_profiletype(
-                profile_jsonc=profile_jsonc, profile_type=profile_type
-            )
-
-        required_profile: dict = self.load_profile_properties(
-            profile_jsonc=profile_jsonc, profile_name=profile_name
-        )
-        service_profile.update(required_profile)
-        return service_profile
-
-    def load(
-        self,
-        profile_name: Union[str, None] = None,
-        profile_type: Union[str, None] = None,
-        profile_args: Union[dict, None] = None,
-    ) -> dict:
-        """Load z/OSMF connection details from a z/OSMF profile.
-
-        Returns
-        -------
-        zosmf_connection
-            z/OSMF connection object
-
-        We will be loading properties from a bottom up fashion,
-        the bottom being the base/default profile properties
-        and the up being the explicitly mentioned Profile.
-
-        Loading Order :
-            Base Profile Properties
-        Overriding Order:
-            Service Profile (profile explicitly mentioned) properties
-
-            User Defined Properties:
-                Base Profile in User Config
-                Service Profile in User Config
-
-            Profile args
-        """
-
-        # load from project config
-        project_profile_props = {}
-        if self._config_dir is None:
-            self.autodiscover_config_dir()
-
-        self._config_filepath = os.path.join(self._config_dir, self._config_filename)
-        project_profile_props = self.load_from_file(
-            filepath=self._config_filepath,
-            profile_name=profile_name,
-            profile_type=profile_type,
-        )
-
-        # load from user config
-        user_profile_props = {}
-        if self._user_config_dir is None:
-            self.autodiscover_user_config_dir()
-
-        if self._user_config_dir:
-            self._user_config_filepath = os.path.join(
-                self._user_config_dir, self._user_config_filename
-            )
-            user_profile_props = self.load_from_file(
-                filepath=self._user_config_filepath,
-                profile_name=profile_name,
-                profile_type=profile_type,
-            )
-
-        project_profile_props.update(user_profile_props)
-
-        return project_profile_props
-
-    def load_credentials(self) -> dict:
-        """
-        return credentials stored for the given config
-        """
-        credentials: dict = {}
+        if not HAS_KEYRING:
+            self.secure_props = {}
 
         try:
             service_name = constants["ZoweServiceName"]
@@ -392,19 +242,128 @@ class ProfileManager:
         # first look for credentials stored for currently loaded config
         # then look for default credential stored for user_directory/.zowe/zowe.config.json
         try:
-            credentials = secure_config_json[self._config_filepath]
+            self.secure_props = secure_config_json[self.location]
         except KeyError:
             try:
-                home = os.path.expanduser("~")
-                global_config_path = os.path.join(home, ".zowe", "zowe.config.json")
-                credentials = secure_config_json[global_config_path]
+                self.secure_props = secure_config_json[GLOBAL_CONFIG_PATH]
             except KeyError as exc:
-                raise Exception(
-                    "No credentials found for loaded config file as well as for global config"
+                raise warnings.warn(
+                    f"No credentials found for loaded {self.filename} file as well as for global config"
                 ) from exc
             else:
                 warnings.warn(
-                    f"Credentials not found for given config, using global credentials {global_config_path}"
+                    f"Credentials not found for given config, using global credentials {GLOBAL_CONFIG_PATH}"
                 )
 
-        return credentials
+
+class ProfileManager:
+    def __init__(self, appname: str = "zowe"):
+        print(CURRENT_DIR)
+        self._appname = appname
+
+        self.project_config = ConfigFile(type=TEAM_CONFIG, name=appname)
+        self.project_user_config = ConfigFile(type=USER_CONFIG, name=appname)
+
+        self.global_config = ConfigFile(type=TEAM_CONFIG, name=GLOBAL_CONFIG_NAME)
+        self.global_user_config = ConfigFile(type=USER_CONFIG, name=GLOBAL_CONFIG_NAME)
+
+    @property
+    def config_appname(self) -> str:
+        """Returns the app name"""
+        return self._appname
+
+    @property
+    def config_dir(self) -> Union[str, None]:
+        """Returns the folder path to where the Zowe z/OSMF Team Project Config files are located."""
+        return self.project_config.directory
+
+    @config_dir.setter
+    def config_dir(self, dirname: str) -> None:
+        """
+        Set directory/folder path to where Zowe z/OSMF Team Project Config files are located
+        """
+        if os.path.isdir(dirname):
+            self.project_config.location = dirname
+        else:
+            raise FileNotFoundError(f"given path {dirname} is not valid")
+
+    @property
+    def user_config_dir(self) -> Union[str, None]:
+        """Returns the folder path to where the Zowe z/OSMF User Project Config files are located."""
+        return self.project_user_config.location
+
+    @user_config_dir.setter
+    def user_config_dir(self, dirname: str) -> None:
+        """Set directory/folder path to where Zowe z/OSMF User Project Config files are located"""
+        if os.path.isdir(dirname):
+            self.project_user_config.location = dirname
+        else:
+            raise FileNotFoundError(f"given path {dirname} is not valid")
+
+    @property
+    def config_filename(self) -> str:
+        """Return the filename for Zowe z/OSMF Team Project Config"""
+        return self.project_config.filename
+
+    @property
+    def config_filepath(self) -> Union[str, None]:
+        """Get the full Zowe z/OSMF Team Project Config filepath"""
+        return self.project_config.filepath
+
+    def load(
+        self,
+        profile_name: Union[str, None] = None,
+        profile_type: Union[str, None] = None,
+    ) -> dict:
+        if profile_name is None and profile_type is None:
+            raise ProfileNotFound(
+                "Could not find profile as both profile_name and profile_type is not set."
+            )
+
+        service_profile: dict = {}
+
+        project_profile: dict = self.project_config.get_profile(
+            profile_name=profile_name, profile_type=profile_type
+        )
+        project_user_profile: dict = self.project_user_config.get_profile(
+            profile_name=profile_name, profile_type=profile_type
+        )
+
+        self.global_config.init_from_file()
+        if profile_name:
+            global_base_profile: dict = self.global_config.load_profile_properties(
+                profile_name=profile_name
+            )
+        else:
+            try:
+                gb_profile_name = self.global_config.get_profilename_from_profiletype(
+                    profile_type=profile_type
+                )
+                global_base_profile: dict = self.global_config.load_profile_properties(
+                    profile_name=gb_profile_name
+                )
+                service_profile.update(global_base_profile)
+            except Exception:
+                warnings.warn("Could not find global base profile")
+
+        self.global_user_config.init_from_file()
+        if profile_name:
+            global_user_profile: dict = self.global_config.load_profile_properties(
+                profile_name=profile_name
+            )
+        else:
+            try:
+                gu_profile_name = self.global_config.get_profilename_from_profiletype(
+                    profile_type=profile_type
+                )
+                global_user_profile: dict = self.global_config.load_profile_properties(
+                    profile_name=gu_profile_name
+                )
+                service_profile.update(global_user_profile)
+            except Exception:
+                warnings.warn("Could not find global user profile")
+
+        service_profile.update(project_profile)
+        service_profile.update(project_user_profile)
+
+        return service_profile
